# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: Oasis API
  description: API routes for Oasis parametric insurance platform
  version: 1.0.0
servers:
  - url: https://your-app.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
paths:
  /api/cron/adjudicator:
    get:
      tags:
        - Cron Jobs
      summary: Run adjudicator
      description: |
        Runs the parametric adjudicator.

        Triggered automatically **every 15 minutes** (GitHub Actions or Supabase pg_cron). Polls weather, AQI, and news APIs for providers that do not support webhooks. For realtime, use POST /api/webhooks/disruption.
      operationId: getCronAdjudicator
      security:
        - cronSecret: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  triggered_at: { type: string, format: date-time, example: "2026-03-06T13:00:00Z" }
                  candidates_found: { type: integer, example: 12 }
                  claims_created: { type: integer, example: 5 }
                  zones_checked: { type: integer, example: 8 }
                  message: { type: string, example: "Hourly processing complete." }
  /api/cron/weekly-premium:
    get:
      tags:
        - Cron Jobs
      summary: Weekly premium renewal
      description: |
        Renews active weekly policies.

        Executes exactly at **Sunday 17:30 UTC** to prepare policies for the upcoming week based on individual zone risk scoring.
      operationId: getCronWeeklyPremium
      security:
        - cronSecret: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  renewed: { type: integer, example: 1250 }
                  skipped: { type: integer, example: 45 }
...
  /api/admin/demo-trigger:
    post:
      tags:
        - Admin Operations
      summary: Inject demo disruption
      description: |
        Inject a synthetic, high-severity disruption event into the database to manually trigger payouts for demos or QA testing.
      operationId: postAdminDemoTrigger
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [eventSubtype, lat, lng]
              properties:
                eventSubtype:
                  type: string
                  enum: [extreme_heat, heavy_rain, severe_aqi, traffic_gridlock, zone_curfew]
                  example: extreme_heat
                lat: { type: number, format: float, example: 12.9716 }
                lng: { type: number, format: float, example: 77.5946 }
                radiusKm: { type: number, default: 15, example: 10 }
                severity: { type: number, example: 9 }
      responses:
        '200':
          description: Synthetic event injected successfully.
  /api/onboarding/verify-face:
    get:
      tags:
        - KYC & Onboarding
      summary: Get face verification gesture
      description: Returns a random liveness gesture (e.g. `smile`, `look_left`) required for the next face verification step.
      operationId: getVerifyFaceGesture
      responses:
        '200':
          description: Random gesture retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  gesture: { type: string, example: "smile" }
    post:
      tags:
        - KYC & Onboarding
      summary: Verify face photo
      description: |
        Validates the user's face photo actively performing the requested gesture.

        This utilizes **OpenRouter Vision Models** to ensure the image satisfies the liveness check criteria before proceeding.
      operationId: postVerifyFace
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [face_photo, expected_gesture]
              properties:
                face_photo: { type: string, format: binary }
                expected_gesture: { type: string, example: "smile" }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  verified: { type: boolean, example: true }
                  face_photo_url: { type: string, example: "https://bucket..." }
                  reason: { type: string, example: "Gesture matched successfully." }
...
components:
  securitySchemes:
    cronSecret:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer CRON_SECRET
    webhookSecret:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer WEBHOOK_SECRET or CRON_SECRET (or X-Webhook-Secret)